% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/admm.R
\name{admm}
\alias{admm}
\title{Performs ADMM to solve FPS}
\usage{
admm(S, ndim, lambda, y0, w0, tau, tauStep = 2, maxiter = 100,
  eps = 10^(-3))
}
\arguments{
\item{S}{Input symmetric matrix}

\item{ndim}{The dimension of Fantope}

\item{lambda}{The smoothing parameter}

\item{maxiter}{The maximum interations}

\item{eps}{Accuracy for the stopping criterion}
}
\value{
H The global solution matrix H for the Fantope problem
}
\description{
Performs ADMM to solve FPS
}

