% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fantopeProj.R
\name{fantopeProj}
\alias{fantopeProj}
\title{A function that projects a matrix to Fantope}
\usage{
fantopeProj(S, ndim)
}
\arguments{
\item{S}{An input p by p symmetric matrix}

\item{ndim}{The dimension of Fantope}
}
\value{
projH The projection matrix of input Smat
}
\description{
A function that projects a matrix to Fantope
}

