% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastADMM-cpp.R
\name{fastADMM_cpp}
\alias{fastADMM_cpp}
\title{Performs ADMM to solve FPS, using fast fantope projection}
\usage{
fastADMM_cpp(S, ndim, lambda, y0, w0, tau, tauStep = 2, maxiter = 100,
  eps = 10^(-3))
}
\arguments{
\item{S}{Input symmetric matrix}

\item{ndim}{The dimension of Fantope}

\item{lambda}{The smoothing parameter}

\item{y0}{Initialized value for aux variable}

\item{w0}{Initialized value for dual variable}

\item{tau}{ADMM penalty}

\item{tauStep}{Step size to adjust tau}

\item{maxiter}{The maximum interations}

\item{eps}{Accuracy for the stopping criterion}
}
\value{
H The global solution matrix H for the Fantope problem
}
\description{
Solve: min_{H, y} infty*I_{Fd}(H) - <S,H> + lambda*|y|_1 + tau/2 |H-y|_F^2
subject to H-y=0
}

