% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastFantopeProj.R
\name{fastFantopeProj}
\alias{fastFantopeProj}
\title{A function that projects a matrix to Fantope -- speed up version}
\usage{
fastFantopeProj(S, ndim)
}
\arguments{
\item{S}{An input p by p symmetric matrix}

\item{ndim}{The dimension of Fantope}
}
\value{
projH The projection matrix of input Smat
}
\description{
A function that projects a matrix to Fantope -- speed up version
}

