% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastFPS.cpp.R
\name{fastFPS.cpp}
\alias{fastFPS.cpp}
\title{The main function that performs fast FPS on a sequence of lambdas}
\usage{
fastFPS.cpp(S, ndim, lambda, maxiter = 100, eps = 0.001, verbose = 0)
}
\arguments{
\item{S}{Input symmetric matrix}

\item{ndim}{The dimension of Fantope}

\item{lambda}{The smoothing parameter, from large to small}

\item{maxiter}{The maximum interations}

\item{eps}{Accuracy for the stopping criterion}

\item{verbose}{how many outputs wanted}
}
\value{
A list that contains the following objects:
}
\description{
The main function that performs fast FPS on a sequence of lambdas
}

