% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/fastFPS.R
\name{fastFPS}
\alias{fastFPS}
\title{The main function that performs fast FPS on a sequence of lambdas}
\usage{
fastFPS(S, ndim, lambdas, maxiter = 100, eps = 0.001, verbose = 0)
}
\arguments{
\item{S}{Input symmetric matrix}

\item{ndim}{The dimension of Fantope}

\item{maxiter}{The maximum interations}

\item{eps}{Accuracy for the stopping criterion}

\item{lambda}{The smoothing parameter}
}
\value{
A list that contains the following objects:
}
\description{
The main function that performs fast FPS on a sequence of lambdas
}

