% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/findActive.R
\name{findActive}
\alias{findActive}
\title{Computes a set of indices that contains the active variables of FPS
If maxoffdiag(i) <= lambda _and_ |{j: diag(j) >= diag(i)}| >= ndim,
then variable i can be safely eliminated.}
\usage{
findActive(S, ndim, lambda)
}
\arguments{
\item{S}{The input symmetric matrix S}

\item{ndim}{The dimension of Fantole}

\item{lambda}{The smoothing parameter}
}
\value{
act_indices, the indices of active set
}
\description{
Computes a set of indices that contains the active variables of FPS
If maxoffdiag(i) <= lambda _and_ |{j: diag(j) >= diag(i)}| >= ndim,
then variable i can be safely eliminated.
}

