% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/getTheta.R
\name{getTheta}
\alias{getTheta}
\title{A function that calculates the threshold theta for Fantope projection}
\usage{
getTheta(v, ndim)
}
\arguments{
\item{v}{A p by 1 vector of sorted eigen values, v_1 >= ... >= v_p}

\item{ndim}{The effective dimension desired}
}
\value{
theta, a threshold such that /sum_i v_i^+ = ndim/,
where x^+ = min(max(x-theta, 0), 1)
}
\description{
A function that calculates the threshold theta for Fantope projection
}

