% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/myFPS.R
\name{myFPS}
\alias{myFPS}
\title{The main function that performs FPS on a sequence of lambdas}
\usage{
myFPS(S, ndim, lambda, maxiter = 100, eps = 0.001, verbose = 0)
}
\arguments{
\item{S}{Input symmetric matrix}

\item{ndim}{The dimension of Fantope}

\item{lambda}{The smoothing parameter, from large to small}

\item{maxiter}{The maximum interations}

\item{eps}{Accuracy for the stopping criterion}

\item{verbose}{How many outputs are wanted}
}
\value{
A list that contains the following objects:
}
\description{
This is an exact replication of Vincent Vu's fps package
}

