% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex}
\alias{simplex}
\title{A function that mimics Vu's simplex function}
\usage{
simplex(v, ndim)
}
\arguments{
\item{v}{A p by 1 vector of sorted eigen values, v_1 >= ... >= v_p}

\item{ndim}{The dimension of simplex}
}
\value{
theta, the threshold
where v^+ = min(max(x-theta, 0), 1)
}
\description{
Should be working the same way as getTheta
}

