% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/simplex.R
\name{simplex_sum}
\alias{simplex_sum}
\title{A function that mimics Vu's simplex_sum function}
\usage{
simplex_sum(v, theta)
}
\arguments{
\item{v}{A p by 1 vector of sorted eigen values, v_1 >= ... >= v_p}

\item{theta}{The threshold}
}
\value{
v_sum, the thresholded sum of vector v^+,
where v^+ = min(max(x-theta, 0), 1)
}
\description{
Calculated the thresholded sum of a vector
}

