% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/softThreshold.R
\name{softThreshold}
\alias{softThreshold}
\title{Performs soft-threshold}
\usage{
softThreshold(x, lambda)
}
\arguments{
\item{x}{A vector or a matrix}

\item{lambda}{The threshold}
}
\value{
xThres, the thresholded x
}
\description{
Performs soft-threshold
}

